Add minion pathing and make actor movement more accurate

+++ game/src/main/java/xyz/openatbp/extension/reqhandlers/MoveActorHandler.java
@@ -75,7 +75,7 @@ public class MoveActorHandler extends BaseClientRequestHandler {
         currentLoc.putFloat("x",(float)movementLine.getX1());
         currentLoc.putFloat("z",(float)movementLine.getY1());
         userLocation.putSFSObject("p1",currentLoc);
-        userLocation.putInt("time",0);
+        userLocation.putFloat("time",0);
         userLocation.putFloat("speed",params.getFloat("speed"));
         ISFSObject data = new SFSObject();
         data.putUtfString("i", String.valueOf(sender.getId()));
